# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DevPort API",
    "description": "DevPort API service",
    "version": "1.1"
  },
  "paths": {
    "/users": {
      "summary": "User endpoint",
      "get": {
        "description": "Returns all users registered on the db.",
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "number"
                    },
                    "user": {
                      "type": "string"
                    },
                    "socials": {
                      "type": "object",
                      "properties": {
                        "twitter": {
                          "type": "string"
                        },
                        "linkedin": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Unsuccessful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Creates a new user.",
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "phone": {
                      "type": "number"
                    },
                    "user": {
                      "type": "string"
                    },
                    "socials": {
                      "type": "object",
                      "properties": {
                        "twitter": {
                          "type": "string"
                        },
                        "linkedin": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}